requirements.txt
streamlit>=1.32.0
openai>=1.0.0
numpy>=1.24.3,<2.0.0
sentence-transformers>=2.2.2
torch>=2.0.1
pymupdf>=1.23.0
python-docx>=0.8.11
transformers>=4.51.3
python-dotenv==1.0.0